stock-analysis/requirements.txt
pandas
yfinance
pyyaml
pyfinviz